requirements.txt
openai
pandas
python-dotenv
loguru
